refactor: hoist universe bible field caps to server/lib/universeBibleLimits.js (#6818)
The six bible-field length caps and the lockable-fields list existed three
times: client/src/services/apiUniverseBuilder.js, server/services/
universeBuilder/sanitize.js, and a third copy of three of the caps in
server/services/pipeline/series.js — with no guard keeping the numbers
aligned. Move them into a new server/lib/universeBibleLimits.js leaf;
sanitize.js and series.js now import from it (restoring the local bindings
their own sanitizers read, since a bare re-export doesn't create one), and
the client re-exports the same leaf under its established WORLD_* names.

Also restores every other apiUniverseBuilder.js export (listUniverses,
createUniverse, style-reference/expand/refine endpoints, the
waitForUniverseWrites/pendingUniverseWrites write-tracking pair, ...) that
had been deleted along with the caps.
