+++ server/services/pipeline/series.js
@@ -18,6 +18,7 @@ import { sanitizeArc, sanitizeSeasonList } from '../../lib/storyArc.js';
 import { sanitizeCharacterArcList } from '../../lib/seriesCharacterArc.js';
 import { sanitizeStyleGuide } from '../../lib/styleGuide.js';
 import { sanitizeProseExportSettings } from '../../lib/proseExportSettings.js';
+import { LOGLINE_MAX, PREMISE_MAX, STYLE_NOTES_MAX } from '../../lib/universeBibleLimits.js';
 import { sanitizeSeverityWeights, sanitizeBlockingSeverities } from '../../lib/editorial/severityConfig.js';
 import { CHECK_SEVERITIES } from '../../lib/editorial/checkInfra/taxonomy.js';
 import { sanitizeOrigin } from '../../lib/sharingOrigin.js';
@@ -53,11 +54,7 @@ const makeErr = (message, code) => Object.assign(new Error(message), { code });
 const SERIES_ID_RE = /^ser-[A-Za-z0-9-]+$/;
 
 export const NAME_MAX = 200;
-export const LOGLINE_MAX = 500;
-// Large enough for a multi-site production bible with concrete route and
-// resource rules; still bounded so the field cannot become an unbounded prompt.
-export const PREMISE_MAX = 20000;
-export const STYLE_NOTES_MAX = 4000;
+export { LOGLINE_MAX, PREMISE_MAX, STYLE_NOTES_MAX };
 // Author-supplied real-world fact reference (#1588) — the ground-truth facts the
 // opt-in `research.fact-accuracy` editorial check reconciles the prose against
 // (e.g. "Paris is the capital of France"; physiological limits). Generous cap so
